added memory hardness and multi-threading
the multi-threading part is a performance parameter, and does not
cause a key change.  so, as you use `ciphart` in different computers
with different counts of cpu cores, you are free to choose the right
number of workers, without being worried to lose your ability to
encrypt/decrypt old files from other machines.