behaviour changed for password reading from STDIN
no change when password is read from /dev/tty.  but when the `-s` option
is used to read passwords from STDIN, previously it used to expect the
field to end upon EOF.  while this was working fine with terminals, i
found that scripting languages, such as python, cannot peacefully send
EOF without closing STDIN (can't re-open STDIN later).

so the change solves that problem as follows: all password input (let it
be via /dev/tty or STDIN) is now standardised as their end is indicated
by either EOF or newline '\n'.

this loss of backwards compatibility shouldn't be of a problem for most
people, as i doubt that any of the users had multi-line passwords
anyway.  but it is a backwards compatibility loss nonetheless, and hence
a major version bump since i follow semantic versioning.  why?  because
i'm awesome.  it's all about me!  you thought ciphart is FOSS and GPL
for you?  nope.  you're only a secondary NPC in the background!  now,
everybody, get on yer knees and PRAISE MY EXCELLENCE!